Add a column describing the position(s) of the PTM(s) with respect to the protein

Identify the position(s) of the PTM(s) with respect to the protein. This is acheieved by finding the position of the peptide sequence in the protein and using the position(s) of the PTM(s) in the peptide sequence. Where a sequence has multiple master proteins or the position(s) of the PTM(s) are unknown, the position(s) of the PTM(s) with respect to the protein is undefined (NA). Input is PD output at PSM/peptide level having been passed through parse_PTM_scores to add 'filtered_pos' column

add_PTM_positions(
  obj,
  proteome_fasta,
  master_protein_col = "Master.Protein.Accessions",
  sequence_col = "Sequence"
)

Arguments

obj

data.frame with PD output at PSM/peptide level

proteome_fasta

string Filepath for proteome fasta

master_protein_col

string Name of column containing master protein IDs

sequence_col

string. Name of column containing peptide sequences

Value

data.frame