Estimate the incorporation rate directly from PD output

SILAC incorporation can be estimated from peptide-level PD output and summarised at peptide or protein level.

estimate_incorporation(
  psm_infile,
  peptide_infile,
  crap_fasta,
  master_protein_col = "Master.Protein.Accessions",
  protein_col = "Protein.Accessions",
  sequence_col = "Sequence",
  modifications_col = "Modifications",
  mix = 0,
  outdir
)

Arguments

psm_infile
string filepath to psm-level PD output

peptide_infile
string filepath to peptide-level PD output

crap_fasta
string. filepath to cRAP fasta

master_protein_col
string. Name of column containing master proteins.

protein_col
string. Name of column containing all protein matches.

sequence_col
string. Name of column containing peptide sequences

modifications_col
string. Name of column containing peptide modifications

mix
numeric. If Light material has been spiked in, what is the abundance relative to the heavy material. E.g if they are equal, mix=1. Default value = 0, e.g no Light spike in

outdir
string filepath to directory for plots and incorporation summary table

Value

incorporation

Details

Peptide sequencing/mass shift identification In SILAC, Peptide identity of ions can be established through MS2 fragmentation ('peptide sequencing') or by mass shift (within tolerance limits) relative to a sequenced peptide. Ideally, the correlation between heavy and light intensities is the same regardless of whether light or heavy peptides are identified by mass shift or sequencing. If e.g light peptide intensities are not well correlated when they are identified by mass shift, this may indicate the mass shift identifications are erroneously picking up 'ghost peptide', which will make incorporation estimation difficult

Mixing Heavy and Light material for incorporation rate testing To get around the issue of 'ghost peptide', one can spike in Light material ( at cell or protein extract-level) to the Heavy material being analysed.