Filter rows based on NA values in column groups

This function is used to filter rows based on number of NA values in different groups of columns. The column groups are defined by a regular expression matching their names. This provides more granular control over what is filtered out compared to filtering based on ratios alone.

The function can be explained as follows:

filter_na(
  data,
  op = "<=",
  setop = dplyr::intersect,
  regex = c("sample_A", "sample_B"),
  value = c(1, 1)
)

The above can be translated as: keep rows that have less than or equal to 1 NA value in the "sample_A" columns AND the "sample_B" columns.

More illustrative examples are below.

filter_na(data, op, setop, regex, value)

Arguments

data

data.frame or matrix to be filtered.

op

string. Defines the operation used for the filtering. Should be one of "==", "<=", ">=", or "!=".

setop

The function used for combining the results for the different column groups. Should be one of dplyr::union (equivalent to "OR"), dplyr::intersect (equivalent to "AND"), or dplyr::setdiff (equivalent to "SYMMETRIC DIFFERENCE").

regex

character vector of length n. The regular expression(s) used to define the column groups. The length of the vector indicates the number of column groups to use.

value

numeric vector of length n. The number of NA values to check for in the rows of each column group. Must be the same length as the regex vector.

Value

Returns a filtered object of the same class as data.

Examples

# set a seed for reproducibility
set.seed(123)

# make a data.frame with fake protein/peptide abundance data
df <- as.data.frame(cbind(replicate(3, runif(7, 13, 24)),
                          replicate(3, runif(7, 24, 30))))

colnames(df) <- c(paste0("ctr", 1:3), paste0("trt", 1:3))

# add in some missing values
mapply(function(i, j) {
  df[i, j] <<- NA_real_
}, c(2:7), list(2, c(2:3), c(3:4), c(2:5), c(2:6), c(1:6)))
#> [1] NA NA NA NA NA NA

# filter for rows with <= 1 NA value in ctr OR trt samples
filter_na(df, "<=", dplyr::union, c("ctr", "trt"), c(1, 1))
#>       ctr1     ctr2     ctr3     trt1     trt2     trt3
#> 1 16.16335 22.81661 14.13217 28.15682 25.73496 26.86678
#> 2 21.67136       NA 22.89807 27.84304 24.88268 28.55076
#> 3 22.71319 23.52517       NA       NA 29.41379 25.90909
#> 4 17.49875       NA       NA 29.96562 29.77815 25.29845

# filter for rows with <= 1 NA value in ctr AND trt samples
filter_na(df, "<=", dplyr::intersect, c("ctr", "trt"), c(1, 1))
#>       ctr1     ctr2     ctr3     trt1     trt2     trt3
#> 1 16.16335 22.81661 14.13217 28.15682 25.73496 26.86678
#> 2 21.67136       NA 22.89807 27.84304 24.88268 28.55076
#> 3 22.71319 23.52517       NA       NA 29.41379 25.90909

# filter for rows with exactly 1 NA value in ctr AND trt samples
filter_na(df, "==", dplyr::intersect, c("ctr", "trt"), c(1, 1))
#>       ctr1     ctr2 ctr3 trt1     trt2     trt3
#> 1 22.71319 23.52517   NA   NA 29.41379 25.90909

# filter for rows with at least 1 non-NA value
filter_na(df, "<=", dplyr::union, "ctr|trt", 5)
#>       ctr1     ctr2     ctr3     trt1     trt2     trt3
#> 1 16.16335 22.81661 14.13217 28.15682 25.73496 26.86678
#> 2 21.67136       NA 22.89807 27.84304 24.88268 28.55076
#> 3 17.49875       NA       NA 29.96562 29.77815 25.29845
#> 4 22.71319 23.52517       NA       NA 29.41379 25.90909
#> 5 23.34514       NA       NA       NA       NA 25.38975
#> 6 13.50112       NA       NA       NA       NA       NA

# filter for rows with at least 3 NA values
filter_na(df, ">=", dplyr::union, "ctr|trt", 3)
#>       ctr1 ctr2 ctr3 trt1 trt2     trt3
#> 5 23.34514   NA   NA   NA   NA 25.38975
#> 6 13.50112   NA   NA   NA   NA       NA
#> 7       NA   NA   NA   NA   NA       NA